+++ frontend/src/Components/NavBar/styles.css
@@ -1,0 +1,47 @@
+
+.Logo {
+    width: 50px;
+    height: auto;
+}
+
+.nav {
+    background-color: #3a584d;
+    display: flex;
+    justify-content: space-between;
+    align-items: stretch;
+    gap: 2rem;
+    padding: 0 1rem;
+}
+
+.site-title {
+    font-size: 2rem;
+}
+
+.nav-bar-container {
+    display: flex;
+    align-items: center;
+    list-style: none;
+    margin: 0;
+    padding: 0;
+}
+
+.nav-bar-item {
+    position: relative;
+    text-decoration: none;
+    margin-right: 20px;
+}
+
+.nav-bar-item a {
+    text-decoration: none;
+    padding: 10px;
+    transition: opacity 0.3s ease-in-out;
+    color: #ff5a0b;
+}
+
+.nav-bar-item a:hover {
+    border-bottom: #ff5a0b solid 3px;
+}
+
+.active {
+    border-bottom: #ff5a0b solid 3px;
+}
